refactor: types in `@/core`
