+++ client-app/App.vue
@@ -18,7 +18,7 @@ import { Head as PageHead } from "@vueuse/head";
 import { markRaw } from "vue";
 import type { Component } from "vue";
 import { useRoute, useRouter } from "vue-router";
-import { useCategories, useDomUtils, useNavigations, useCatalogMenu } from "@/core";
+import { useCategories, useDomUtils, useNavigations, useCatalogMenu } from "@/core/composables";
 import { useCart } from "@/shared/cart";
 import { NotificationsHost } from "@/shared/notification";
 import { PopupHost } from "@/shared/popup";
