--- client-app/core/index.ts
@@ -1,10 +1,0 @@
-export * from "./composables";
-export * from "./directives";
-export * from "./enums";
-export * from "./plugins";
-export * from "./types";
-export * from "./utilities";
-
-export * from "./constants";
-export * from "./injection-keys";
-export * from "./security-constants";
