+++ client-app/pages/account/lists.vue
@@ -70,7 +70,8 @@
 import { breakpointsTailwind, useBreakpoints } from "@vueuse/core";
 import { computed, inject } from "vue";
 import { useI18n } from "vue-i18n";
-import { DEFAULT_WISHLIST_LIMIT, usePageHead } from "@/core";
+import { usePageHead } from "@/core/composables";
+import { DEFAULT_WISHLIST_LIMIT } from "@/core/constants";
 import { configInjectionKey } from "@/core/injection-keys";
 import { usePopup } from "@/shared/popup";
 import {
