+++ client-app/pages/account/orders.vue
@@ -302,8 +302,8 @@ import { breakpointsTailwind, useBreakpoints, onClickOutside } from "@vueuse/cor
 import { computed, onMounted, ref, shallowRef, watch } from "vue";
 import { useI18n } from "vue-i18n";
 import { useRouter } from "vue-router";
-import { ISortInfo } from "@/core";
 import { useElementVisibility, usePageHead } from "@/core/composables";
+import type { ISortInfo } from "@/core/types";
 import {
   OrdersFilter,
   MobileOrdersFilter,
@@ -311,7 +311,7 @@ import {
   useUserOrders,
   PageToolbarBlock,
 } from "@/shared/account";
-import { CustomerOrderType } from "@/xapi/types";
+import type { CustomerOrderType } from "@/xapi/types";
 
 const { t } = useI18n();
 const router = useRouter();
