+++ client-app/pages/company/info.vue
@@ -239,12 +239,12 @@ import { useI18n } from "vue-i18n";
 import * as yup from "yup";
 import { usePageHead } from "@/core/composables";
 import { AddressType, XApiPermissions } from "@/core/enums";
-import { ISortInfo } from "@/core/types";
+import type { ISortInfo } from "@/core/types";
 import { useUser } from "@/shared/account";
 import { AddOrUpdateCompanyAddressDialog, useOrganization, useOrganizationAddresses } from "@/shared/company";
 import { useNotifications } from "@/shared/notification";
 import { usePopup } from "@/shared/popup";
-import { MemberAddressType } from "@/xapi";
+import type { MemberAddressType } from "@/xapi/types";
 
 const page = ref(1);
 const itemsPerPage = ref(10);
