+++ client-app/shared/account/components/order-line-items.vue
@@ -156,8 +156,9 @@
 <script setup lang="ts">
 import { sumBy } from "lodash";
 import { computed } from "vue";
-import { ExtendedLineItemType, extendLineItem } from "@/core";
-import { LineItemType, OrderLineItemType } from "@/xapi";
+import type { ExtendedLineItemType } from "@/core/types";
+import { extendLineItem } from "@/core/utilities";
+import type { LineItemType, OrderLineItemType } from "@/xapi/types";
 
 interface IProps {
   items?: (OrderLineItemType | LineItemType)[];
