+++ client-app/shared/account/composables/useUserQuotes.ts
@@ -1,6 +1,6 @@
 import { Ref, ref, shallowRef } from "vue";
-import { defaultSortInfo, ISortInfo } from "@/core";
-import { Logger, getSortingExpression } from "@/core/utilities";
+import type { ISortInfo } from "@/core/types";
+import { defaultSortInfo, getSortingExpression, Logger } from "@/core/utilities";
 import { getQuotes } from "@/xapi/graphql/quotes";
 import { QuoteType } from "@/xapi/types";
