+++ client-app/shared/cart/components/cart-line-item.vue
@@ -147,9 +147,10 @@
 </template>
 
 <script setup lang="ts">
-import { ref, computed, watchEffect } from "vue";
-import { ExtendedLineItemType, extendLineItem } from "@/core";
-import { LineItemType } from "@/xapi";
+import { computed, ref, watchEffect } from "vue";
+import type { ExtendedLineItemType } from "@/core/types";
+import { extendLineItem } from "@/core/utilities";
+import type { LineItemType } from "@/xapi/types";
 
 interface IProps {
   disabled?: boolean;
