+++ client-app/shared/layout/components/header/_internal/mobile-menu-link.vue
@@ -47,21 +47,25 @@
 
 <script setup lang="ts">
 import { computed } from "vue";
-import { NavigationFailure } from "vue-router";
-import { MenuLinkType, numberToShortString } from "@/core";
+import type { NavigationFailure } from "vue-router";
+import type { MenuLinkType } from "@/core/types";
+import { numberToShortString } from "@/core/utilities";
 
-interface Props {
-  link: MenuLinkType;
-  count?: number;
-}
-
-interface Emits {
+interface IEmits {
   (event: "select"): void;
   (event: "close"): void;
 }
 
-const emit = defineEmits<Emits>();
-const props = withDefaults(defineProps<Props>(), { count: 0 });
+interface IProps {
+  link: MenuLinkType;
+  count?: number;
+}
+
+const emit = defineEmits<IEmits>();
+const props = withDefaults(defineProps<IProps>(), {
+  count: 0,
+});
+
 const isParent = computed<boolean>(() => !!props.link.children?.length);
 const isLink = computed<boolean>(() => !!props.link.route && !isParent.value);
 const preparedCount = computed<string>(() => numberToShortString(props.count));
