+++ client-app/ui-kit/components/molecules/address-selection/vc-address-selection.vue
@@ -32,7 +32,7 @@
 </template>
 
 <script setup lang="ts">
-import { AnyAddressType } from "@/core";
+import type { AnyAddressType } from "@/core/types";
 
 interface IEmits {
   (event: "change"): void;
